+++ Sources/Views/Cells/MessageCollectionViewCell.swift
@@ -25,12 +25,7 @@
 import UIKit
 
 /// A subclass of `UICollectionViewCell` to be used inside of a `MessagesCollectionView`.
-open class MessageCollectionViewCell: UICollectionViewCell, CollectionViewReusable {
-
-    /// The reuse identifier of the cell.
-    open class func reuseIdentifier() -> String {
-        return "messagekit.cell.base-cell"
-    }
+open class MessageCollectionViewCell: UICollectionViewCell {
 
     // MARK: - Initializers